refactor(#325) - xml , 어뎁터 , 안쓰는 파일 제거
