--- AOS/app/src/main/java/boostcamp/and07/mindsync/ui/boardlist/BoardClickListener.kt
@@ -1,9 +1,0 @@
-package boostcamp.and07.mindsync.ui.boardlist
-
-import boostcamp.and07.mindsync.data.model.Board
-
-interface BoardClickListener {
-    fun onClick(board: Board)
-
-    fun onCheckBoxClick(board: Board)
-}
