--- AOS/app/src/main/res/layout/fragment_board_list.xml
@@ -1,72 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
-
-    <data>
-
-        <variable
-            name="vm"
-            type="boostcamp.and07.mindsync.ui.boardlist.BoardListViewModel" />
-
-        <import type="com.google.android.flexbox.FlexDirection"></import>
-    </data>
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        tools:context=".ui.boardlist.BoardListFragment">
-
-        <androidx.constraintlayout.widget.Guideline
-            android:id="@+id/gl_board_list_start"
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"
-            android:orientation="vertical"
-            app:layout_constraintGuide_percent="0.1" />
-
-
-        <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/rv_board_list_board"
-            android:layout_width="0dp"
-            android:layout_height="0dp"
-            android:layout_gravity="center|top"
-            android:background="@android:color/transparent"
-            app:boards="@{vm.boardUiState.boards}"
-            app:flexBoxLayoutManager="@{FlexDirection.ROW}"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="@id/gl_board_list_end"
-            app:layout_constraintStart_toEndOf="@id/gl_board_list_start"
-            app:layout_constraintTop_toTopOf="parent"
-            tools:listitem="@layout/item_board" />
-
-        <com.google.android.material.floatingactionbutton.FloatingActionButton
-            android:id="@+id/btn_board_list_refresh"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_margin="10dp"
-            android:background="@android:color/transparent"
-            android:src="@drawable/ic_refresh_board"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:rippleColor="@color/mindmap2" />
-
-
-        <com.google.android.material.floatingactionbutton.FloatingActionButton
-            android:id="@+id/btn_board_list_add_board"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_margin="10dp"
-            app:floatingButtonImage="@{vm.boardUiState.selectBoards}"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:rippleColor="@color/mindmap2" />
-
-        <androidx.constraintlayout.widget.Guideline
-            android:id="@+id/gl_board_list_end"
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"
-            android:orientation="vertical"
-            app:layout_constraintGuide_percent="0.9" />
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-</layout>
