--- AOS/app/src/main/res/layout/item_board.xml
@@ -1,70 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
-
-    <data>
-
-        <variable
-            name="board"
-            type="boostcamp.and07.mindsync.data.model.Board" />
-    </data>
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:padding="5dp">
-
-        <androidx.cardview.widget.CardView
-            android:id="@+id/cv_board_thumbnail"
-            android:layout_width="100dp"
-            android:layout_height="100dp"
-            app:layout_constraintStart_toEndOf="@id/cb_board"
-            app:layout_constraintTop_toTopOf="parent"
-            >
-
-            <ImageButton
-                android:id="@+id/imgbtn_board_item"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:background="@null"
-                android:scaleType="centerCrop"
-                app:boardImageUri="@{board.imageUrl}"
-                android:src="@mipmap/ic_app_logo" />
-        </androidx.cardview.widget.CardView>
-
-        <CheckBox
-            android:id="@+id/cb_board"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            />
-
-        <TextView
-            android:id="@+id/tv_board_title"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:contentDescription="@string/board_list_board_name"
-            android:fontFamily="@font/pretendard_bold"
-            android:text="@{board.name}"
-            android:singleLine="true"
-            android:textAlignment="center"
-            app:layout_constraintEnd_toEndOf="@id/cv_board_thumbnail"
-            app:layout_constraintStart_toStartOf="@id/cv_board_thumbnail"
-            app:layout_constraintTop_toBottomOf="@id/cv_board_thumbnail"
-            tools:text="@string/board_list_board_name" />
-
-        <TextView
-            android:id="@+id/tv_board_date"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:contentDescription="@string/board_list_board_create_date"
-            android:fontFamily="@font/pretendard_bold"
-            app:date="@{board.date}"
-            app:layout_constraintEnd_toEndOf="@id/cv_board_thumbnail"
-            app:layout_constraintStart_toStartOf="@id/cv_board_thumbnail"
-            app:layout_constraintTop_toBottomOf="@id/tv_board_title"
-            tools:text="@string/board_list_board_create_date" />
-    </androidx.constraintlayout.widget.ConstraintLayout>
-</layout>
